Move `TimeToAppointment` to the `overdue` package
This is in preparation for a future feature where we will use this
class in the appointment reminder sheet.
