+++ app/src/main/java/org/simple/clinic/overdue/AppointmentConfig.kt
@@ -1,6 +1,5 @@
 package org.simple.clinic.overdue
 
-import org.simple.clinic.scheduleappointment.TimeToAppointment
 import org.threeten.bp.Period
 
 data class AppointmentConfig(
