+++ app/src/main/java/org/simple/clinic/scheduleappointment/ScheduleAppointmentSheetController.kt
@@ -14,11 +14,12 @@ import org.simple.clinic.overdue.Appointment.AppointmentType.Automatic
 import org.simple.clinic.overdue.Appointment.AppointmentType.Manual
 import org.simple.clinic.overdue.AppointmentConfig
 import org.simple.clinic.overdue.AppointmentRepository
+import org.simple.clinic.overdue.TimeToAppointment
 import org.simple.clinic.patient.PatientRepository
 import org.simple.clinic.protocol.ProtocolRepository
-import org.simple.clinic.scheduleappointment.TimeToAppointment.Days
-import org.simple.clinic.scheduleappointment.TimeToAppointment.Months
-import org.simple.clinic.scheduleappointment.TimeToAppointment.Weeks
+import org.simple.clinic.overdue.TimeToAppointment.Days
+import org.simple.clinic.overdue.TimeToAppointment.Months
+import org.simple.clinic.overdue.TimeToAppointment.Weeks
 import org.simple.clinic.user.UserSession
 import org.simple.clinic.util.UserClock
 import org.simple.clinic.util.filterAndUnwrapJust
